+++ CHANGELOG.md
@@ -3,12 +3,14 @@
 Mapbox welcomes participation and contributions from everyone.
 
 # main
-
 ## Features ✨ and improvements 🏁
 * [compose] Mark `MapState`, `MapViewportState`, `TerrainState`, light states and source states as `Stable` as they are internally backed by `MutableState`.
 * [compose] Add more config options including `showPlaceLabels`, `showRoadLabels`, `showPointOfInterestLabels`, `showTransitLabels` and `font` to `MapboxStandardStyle`.
-* Modified `awaitCameraForCoordinates` extension function to use `MapCameraManagerDelegate` as receiver type.
-* Modified `queryRenderedFeatures` and `querySourceFeatures` extension functions to use `MapFeatureQueryDelegate` as receiver type.
+* Modify `awaitCameraForCoordinates` extension function to use `MapCameraManagerDelegate` as receiver type.
+* Modify `queryRenderedFeatures` and `querySourceFeatures` extension functions to use `MapFeatureQueryDelegate` as receiver type.
+* Introduce asynchronous overloaded method `ViewAnnotationManager.cameraForAnnotations` better covering some corner cases.
+* Make use of asynchronous `MapboxMap.cameraForCoordinates` in Mapbox overlay plugin better covering some corner cases.
+* Mark synchronous methods `MapboxMap.cameraForCoordinates` and `ViewAnnotationManager.cameraForAnnotations` with `@MapboxDelicateApi`. Consider using asynchronous overloaded methods instead.
 
 # 11.5.0 July 05, 2024
 ## Breaking changes ⚠️